Update prototyping.md
I've played around with the order so it follows what I think might be the thought flow - see what you think. And I added a bit of colour to the language too - see if you think it's appropriate / in keeping with the tone of the rest of the site

a few questions:

what exactly is A commissioning organisation looking to increase in-house expertise?

Is it definitely up to a week and not up to two weeks?

I don't like saying that they won't get a "finished product" - is there a way we can manage expectations without implying that the output is at all deficient? Will have a think